feat: second argument as option hash instead of directly the locale
The optional second argument of the parse functions (first argument in `dayjsNow()`) was the locale.
Now it is an object with the options, the first one of which is the `locale`. This is a breaking
change.

BREAKING CHANGE: The optional argument of the custom locale has been replaced by an optional object
with all the options. The first one of which is locale. e.g. `parseDayjs(d, customLocale)` =>
`parseDayjs(d, { locale: customLocale })`