feat: second argument as option hash instead of directly the origin (parseFromStandardPeriods)
the optional second arguemnt of the parseFromStandardPeriods function was the origin, now it is an
object with the options, the first of which is the `origin`

BREAKING CHANGE: The optional argument of the custom origin has been replaced by an optional object
with all the options. The first option available is the `origin`. e.g.
`parseFromStandardPeriods("1d", customOrigin)` => `parseFromStandardPeriods("1d", { origin:
customOrigin })`